Spring security task done

+++ src/main/java/com/epam/gym/controllers/TraineesController.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,6 +23,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.epam.gym.dto.AuthenticationResponse;
+import com.epam.gym.dto.AuthenticationDTO;
 import com.epam.gym.dto.TrainingDTO;
 import com.epam.gym.dto.UserCredentialsDTO;
 import com.epam.gym.dto.trainee.TraineeCreateDTO;
@@ -35,65 +39,82 @@ import com.epam.gym.entities.Training;
 import com.epam.gym.entities.TrainingType;
 import com.epam.gym.entities.TrainingTypeEnum;
 import com.epam.gym.exceptions.ResourceNotFoundException;
+import com.epam.gym.exceptions.AccessDeniedException;
 import com.epam.gym.repositories.TrainingRepository;
+import com.epam.gym.services.AuthenticationService;
 import com.epam.gym.services.TraineeService;
 import com.epam.gym.services.TrainingService;
 
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 
 import com.epam.gym.payloads.ApiResponse;
 
 
 @RestController
 @RequestMapping("/api/v1/trainees")
+@RequiredArgsConstructor
 public class TraineesController {
 
 	private final TraineeService traineeService;
 	private final TrainingService trainingService;
-	
-	@Autowired
-	public TraineesController(TraineeService traineeService,TrainingService trainingService) {
-		this.traineeService=traineeService;
-		this.trainingService=trainingService;
-	}
-	
+	private final AuthenticationService auhtService;
+
 	@GetMapping("/{username}")
-	public ResponseEntity<TraineeDTO> getTraineeByUsername(@PathVariable String username) throws ResourceNotFoundException{
+	public ResponseEntity<TraineeDTO> getTraineeByUsername(@PathVariable String username) throws ResourceNotFoundException,AccessDeniedException{
+		if(!auhtService.hasPermission(username, "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to get profile of user",username);
+		}
 		return new ResponseEntity<>(traineeService.getTraineeProfile(username), HttpStatus.OK) ;
 	}
 	
 	@PutMapping("/{username}")
-	public ResponseEntity<TraineeDTO> updateTraineeProfile(@PathVariable String username,@Valid @RequestBody TraineeUpdateDTO traineeUpdateDTO) throws ResourceNotFoundException{
+	public ResponseEntity<TraineeDTO> updateTraineeProfile(@PathVariable String username,@Valid @RequestBody TraineeUpdateDTO traineeUpdateDTO) throws ResourceNotFoundException,AccessDeniedException{
+		if(!auhtService.hasPermission(username, "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to update profile of user",username);
+		}
 		TraineeDTO traineeUpdated=this.traineeService.updateTraineeProfile(username, traineeUpdateDTO);
-		return new ResponseEntity<>(traineeUpdated, HttpStatus.CREATED) ;
+		return new ResponseEntity<>(traineeUpdated, HttpStatus.CREATED);}
+	
+	
+	@PostMapping("/auth/register")
+	public ResponseEntity<UserCredentialsDTO> registerTrainee(@Valid @RequestBody TraineeCreateDTO traineeCreateDTO){
+		UserCredentialsDTO authenticationResponse=this.traineeService.registerNewTrainee(traineeCreateDTO);
+		return new ResponseEntity<>(authenticationResponse, HttpStatus.CREATED);
 	}
 	
-	@PostMapping("/register")
-	public ResponseEntity<UserCredentialsDTO> registerUser(@Valid @RequestBody TraineeCreateDTO traineeCreateDTO){
-		UserCredentialsDTO userCredentialsDTO=this.traineeService.registerNewTrainee(traineeCreateDTO);
-		return new ResponseEntity<>(userCredentialsDTO, HttpStatus.CREATED);
-	}
-
 	
 	@DeleteMapping("/{username}")
-	public ApiResponse deletePost(@PathVariable String username) throws ResourceNotFoundException {
+	public ApiResponse deleteTrarinee(@PathVariable String username) throws ResourceNotFoundException,AccessDeniedException {
+		if(!auhtService.hasPermission(username, "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to delete profile of user",username);
+		}
 		this.traineeService.deleteTrainee(username);
 		return new ApiResponse("Username is successfully deleted !!", true);
 	}
 	
 	@PutMapping("/{username}/trainers")
-	public ResponseEntity<List<TraineeTrainerDTO>> updateTraineTrainers(@PathVariable String username, @Valid @RequestBody TraineeUpdateTrainersDTO traineeUpdateTrainersDTO)throws ResourceNotFoundException{
+	public ResponseEntity<List<TraineeTrainerDTO>> updateTraineTrainers(@PathVariable String username, @Valid @RequestBody TraineeUpdateTrainersDTO traineeUpdateTrainersDTO)throws ResourceNotFoundException,AccessDeniedException{
+		if(!auhtService.hasPermission(username, "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to update trainers of user",username);
+		}
 		List<TraineeTrainerDTO> traineesTrainersUpdated=this.traineeService.updateTraineTrainers(username, traineeUpdateTrainersDTO);
 		return new ResponseEntity<>(traineesTrainersUpdated, HttpStatus.OK) ;
 	}
 	
 	@GetMapping("/{username}/trainings")
-	public ResponseEntity<List<Object[]>> getTraineeTrainingList(@PathVariable String username, @Valid @ModelAttribute TrainingsFilterDTO trainingsFilterDTO)throws ResourceNotFoundException { 
+	public ResponseEntity<List<Object[]>> getTraineeTrainingList(@PathVariable String username, @Valid @ModelAttribute TrainingsFilterDTO trainingsFilterDTO)throws ResourceNotFoundException,AccessDeniedException { 
+		if(!auhtService.hasPermission(username, "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to get trainings list of user",username);
+		}
 		return new ResponseEntity<>(traineeService.findTraineeTrainingList(username,trainingsFilterDTO),HttpStatus.OK);
 	}
 
 	@PostMapping("/trainings")
-	public ApiResponse saveTraineeTrainings(@Valid @RequestBody TrainingDTO trainingDTO)throws ResourceNotFoundException{
+	public ApiResponse saveTraineeTrainings(@Valid @RequestBody TrainingDTO trainingDTO)throws ResourceNotFoundException,AccessDeniedException{
+		if(!auhtService.hasPermission(trainingDTO.getTraineeUsername(), "TRAINEE")) {
+			throw new AccessDeniedException(auhtService.actualUsername(),"to update trainings list of user",trainingDTO.getTraineeUsername());
+		}
 		this.trainingService.saveTraining(trainingDTO);
 		return new ApiResponse("Training saved sucesfully !!", true);
 	}
